pyproject.toml
[project]
name = "macropy"
version = "0.1.0"
description = "A python macro script"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "python-dotenv>=1.2.1",
]
